+++ packages/wallet/ui/public/bridge.html
@@ -8,7 +8,7 @@
 </head>
 
 <body style="overflow-x: hidden; overflow-y: scroll">
-    <noscript>Without JavaScript enabled, this page cannot function.</noscript>
+    <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
 </body>
